requirements.txt
# ── Data & Versioning ────────────────────────────────────────
dvc>=3.0.0
pandas>=2.0.0
polars>=0.20.0
pyarrow>=15.0.0

# ── Vector Store & Embeddings ─────────────────────────────────
chromadb>=0.5.0
sentence-transformers>=3.0.0

# ── Local LLM via Ollama ──────────────────────────────────────
ollama>=0.3.0
langchain-ollama>=0.1.0

# ── MCP ───────────────────────────────────────────────────────
fastmcp>=2.0.0

# ── Agent Framework ───────────────────────────────────────────
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0

# ── Observability ─────────────────────────────────────────────
langfuse>=2.36.0

# ── Agent Evaluation ─────────────────────────────────────────
deepeval>=1.0.0

# ── Fine-tuning (optional — install separately if needed) ────
# pip install "unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git"
transformers>=4.40.0
datasets>=2.20.0
peft>=0.12.0
trl>=0.9.0
accelerate>=0.33.0

# ── Serving ───────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
httpx>=0.27.0

# ── Orchestration ─────────────────────────────────────────────
prefect>=3.0.0

# ── Monitoring metrics ────────────────────────────────────────
prometheus-client>=0.20.0

# ── Config management ─────────────────────────────────────────
python-dotenv>=1.0.0

# ── Testing ───────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
